+++ resources/src/babashka/babashka/impl/mvn/settings.clj
@@ -19,10 +19,32 @@
                        (System/getProperty key)
                        whole)))))
 
-(defn- server [el]
-  [(child-text el "id")
-   {:username (interpolate (child-text el "username"))
-    :password (interpolate (child-text el "password"))}])
+(defn- server
+  "Returns [id credentials], with :headers from configuration/httpHeaders."
+  [el]
+  (let [id (child-text el "id")
+        headers (reduce
+                 (fn [acc p]
+                   (let [name-el (child p "name")
+                         value-el (child p "value")]
+                     (when-not (and name-el value-el)
+                       ;; the JVM tools.deps fails to start when either element is missing
+                       (throw (ex-info (str "Invalid httpHeaders property for server " id
+                                            " in settings.xml: name and value are required")
+                                       {:server id})))
+                     (let [header-name (or (interpolate (text name-el)) "")
+                           value (or (interpolate (text value-el)) "")]
+                       ;; skip a nameless header, and keep the last of names that differ only in case
+                       (if (= "" header-name)
+                         acc
+                         (assoc (into {} (remove #(.equalsIgnoreCase header-name ^String (key %))) acc)
+                                header-name value)))))
+                 {}
+                 (some-> (child el "configuration") (child "httpHeaders") (children "property")))]
+    [id
+     (cond-> {:username (interpolate (child-text el "username"))
+              :password (interpolate (child-text el "password"))}
+       (seq headers) (assoc :headers headers))]))
 
 (defn- mirror [el]
   {:id (child-text el "id")
@@ -55,8 +77,7 @@
   "Settings from an XML string."
   [s]
   (let [root (babashka.impl.mvn.xml/parse s)]
-    {:local-repository (interpolate (child-text root "localRepository"))
-     :servers (into {} (map server (some-> (child root "servers") (children "server"))))
+    {:servers (into {} (map server (some-> (child root "servers") (children "server"))))
      :mirrors (mapv mirror (some-> (child root "mirrors") (children "mirror")))
      :proxies (mapv proxy-entry (some-> (child root "proxies") (children "proxy")))
      :profiles (into {} (map profile (some-> (child root "profiles") (children "profile"))))